requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
blis==0.7.9
boto3==1.34.56
botocore==1.34.56
catalogue==2.0.8
certifi==2023.11.17
click==8.1.3
cmake==3.25.0
confection==0.0.4
contourpy==1.0.7
cudf==23.4.1
cycler==0.11.0
cymem==2.0.7
Cython==0.29.35
datasets==2.16.1
dill==0.3.7
fastai==2.7.12
fastai2==0.0.30
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.9.0
fonttools==4.39.4
frozenlist==1.4.1
fsspec==2023.5.0
hdbscan==0.8.37
hpack==4.0.0
imageio==2.28.1
importlib_metadata==8.0.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
lit==15.0.7
llvmlite==0.39.1
MarkupSafe==2.1.3
matplotlib==3.7.1
mpi4py==3.1.5
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.15
munkres==1.1.4
murmurhash==1.0.9
pandas==1.4.3
pathy==0.10.1
Pillow==9.5.0
preshed==3.0.8
protobuf==4.21.12
pyarrow==10.0.1
pycocotools==2.0.7
pycw==0.0.3
pydantic==1.10.7
PyLaTeX==1.4.2
pylibraft==23.4.1
pylocron==0.2.0
pynndescent==0.5.13
pyparsing==3.0.9
PyWavelets==1.4.1
requests==2.31.0
rmm==23.4.1
s3transfer==0.10.0
scipy==1.9.1
sewar==0.4.6
spacy==3.5.2
srsly==2.4.6
sympy==1.11.1
tabulate==0.9.0
thinc==8.1.10
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==2.0.1
torchaudio==2.0.2
torchdata==0.7.1
torchmetrics==1.5.2
torchnet==0.0.4
torchvision==0.15.2
tqdm==4.65.0
treelite==3.2.0
triton==2.0.0
typer==0.7.0
umap==0.1.1
urllib3==1.26.18
venn==0.1.3
visdom==0.2.4
wasabi==1.1.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
zstandard==0.23.0
opencv-python==4.12.0.88
numpy==1.23.5
randaugment==1.0.2
easydict==1.13
insightface==0.7.3
onnxruntime-gpu==1.15.1
